sports-backend/requirements.txt
requests==2.31.0
python-dotenv==1.0.0
psycopg2-binary==2.9.9
Flask==3.0.0
gunicorn==21.2.0
python-dateutil==2.8.2
